Early Build
This was any early build of the website while we were still designing
it. Some of the main differences are:
Providers & Team
Services
Missing Nav Sections
More of a basic site that we presented to the admin staff on 06/22/16.
This is mostly here just to see the transformation into what it is now.