Refactoring
There were a lot of changes prior to going live. This resulted in some
sloppy code put into production that worked but was not well structured
or easy to follow in some places. I spent about a week and a half
organizing the css and html files. I need to move on to the apps so this
is all for now. Some major changes:
*Organized CSS
*HTML5 type of tags(header, main, footer) used in substitution of the
three previous divs to making searching through the code easier.
*Spacing should now be consistent in both html and css.
*Header tags are now in place instead of a lot of p tags
*Deleted almost all instances of the br tag where it was not used
semantically.
*Piwik script added to archive sections
*Some images were cleaned out and renamed to make more sense.