Locations Index Fixed
The locations index page had an issue with not clearing the footer. The
issue has been fixed and the titles for each location have been changed
to the locations name for users to better identify them. There was also
an issue with the slideshow after going live with the app landing page,
some of the slideshow code was overwritten. This has been fixed.