Providers Page Clean Up & Partially Done
*Providers page has been cleaned up. The parent divs for each team
member contained too many child divs and it was difficult to
understand/make changes too. So it has been reduced until bios and pics
are added.
*The vpchc.tab.js file was added which contains the code for toggling
the tabs. I may combine this in the future with another script.
*JQueryRotate.js was moved to the scripts folder. I may do this with all
of the scripts in the js_serve folder.
*Part of the transition is done. This is just a backup in case anything
happens over the weekend.